requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.12.0
transformers>=4.35.2
torch>=2.6.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.2
textblob>=0.17.1
pydantic>=2.5.0
python-multipart>=0.0.6
